requirements.txt
pandas
python-docx-template
python-docx
pywin32